requirements.txt
camelot-py[cv]
pandas
pdf2image
pillow
pytesseract
